requirements.txt
colorama
windows-curses
pycryptodome